Initial CVL common infra changes
