+++ cvl/cvl.go
@@ -25,7 +25,7 @@ import (
 	"regexp"
 	"time"
 	 log "github.com/golang/glog"
-	"encoding/json"
+	//"encoding/json"
 	"github.com/go-redis/redis"
 	"github.com/antchfx/xmlquery"
 	"github.com/antchfx/jsonquery"
@@ -33,7 +33,7 @@ import (
 	. "github.com/Azure/sonic-mgmt-common/cvl/internal/util"
 	"sync"
 	"flag"
-	"runtime"
+	//"runtime"
 )
 
 //DB number 
@@ -557,14 +557,6 @@ func getRedisToYangKeys(tableName string, redisKey string)[]keyValuePairStruct{
 	return mkeys
 }
 
-
-//Add child node to a parent node
-func(c *CVL) addChildNode(tableName string, parent *yparser.YParserNode, name string) *yparser.YParserNode {
-
-	//return C.lyd_new(parent, modelInfo.tableInfo[tableName].module, C.CString(name))
-	return c.yp.AddChildNode(modelInfo.tableInfo[tableName].module, parent, name)
-}
-
 //Check for path resolution
 func (c *CVL) checkPathForTableEntry(tableName string, currentValue string, cfgData *CVLEditConfigData, mustExpStk []string, token string) ([]string, string, CVLRetCode) {
 
@@ -929,54 +921,6 @@ func (c *CVL) addTableEntryToCache(tableName string, redisKey string) {
 	}
 }
 
-//Check delete constraint for leafref if key/field is deleted
-func (c *CVL) checkDeleteConstraint(cfgData []CVLEditConfigData,
-			tableName, keyVal, field string) CVLRetCode {
-	var leafRefs []tblFieldPair
-	if (field != "") {
-		//Leaf or field is getting deleted
-		leafRefs = c.findUsedAsLeafRef(tableName, field)
-	} else {
-		//Entire entry is getting deleted
-		leafRefs = c.findUsedAsLeafRef(tableName, modelInfo.tableInfo[tableName].keys[0])
-	}
-
-	//The entry getting deleted might have been referred from multiple tables
-	//Return failure if at-least one table is using this entry
-	for _, leafRef := range leafRefs {
-		TRACE_LOG(INFO_API, (TRACE_DELETE | TRACE_SEMANTIC), "Checking delete constraint for leafRef %s/%s", leafRef.tableName, leafRef.field)
-		//Check in dependent data first, if the referred entry is already deleted
-		leafRefDeleted := false
-		for _, cfgDataItem := range cfgData {
-			if (cfgDataItem.VType == VALIDATE_NONE) &&
-			(cfgDataItem.VOp == OP_DELETE ) &&
-			(strings.HasPrefix(cfgDataItem.Key, (leafRef.tableName + modelInfo.tableInfo[leafRef.tableName].redisKeyDelim + keyVal + modelInfo.tableInfo[leafRef.tableName].redisKeyDelim))) {
-				//Currently, checking for one entry is being deleted in same session
-				//We should check for all entries
-				leafRefDeleted = true
-				break
-			}
-		}
-
-		if (leafRefDeleted == true) {
-			continue //check next leafref
-		}
-
-		//Else, check if any referred enrty is present in DB
-		var nokey []string
-		refKeyVal, err := luaScripts["find_key"].Run(redisClient, nokey, leafRef.tableName,
-		modelInfo.tableInfo[leafRef.tableName].redisKeyDelim, leafRef.field, keyVal).Result()
-		if (err == nil &&  refKeyVal != "") {
-			CVL_LOG(ERROR, "Delete will violate the constraint as entry %s is referred in %s", tableName, refKeyVal)
-
-			return CVL_SEMANTIC_ERROR
-		}
-	}
-
-
-	return CVL_SUCCESS
-}
-
 //Add the data which are referring this key
 func (c *CVL) updateDeleteDataToCache(tableName string, redisKey string) {
 	if _, existing := c.tmpDbCache[tableName]; existing == false {
@@ -1088,24 +1032,6 @@ func (c *CVL) addLeafRef(config bool, tableName string, name string, value strin
 	}
 }
 
-
-func (c *CVL) addChildLeaf(config bool, tableName string, parent *yparser.YParserNode, name string, value string) {
-
-	/* If there is no value then assign default space string. */
-	if len(value) == 0 {
-                value = " "
-        }
-
-	//Batch leaf creation
-	c.batchLeaf = c.batchLeaf + name + "#" + value + "#"
-	//Check if this leaf has leafref,
-	//If so add the add redis key to its table so that those 
-	// details can be fetched for dependency validation
-
-	c.addLeafRef(config, tableName, name, value)
-}
-
-
 func (c *CVL) checkFieldMap(fieldMap *map[string]string) map[string]interface{} {
 	fieldMapNew := map[string]interface{}{}
 
@@ -1136,351 +1062,6 @@ func mergeMap(dest map[string]string, src map[string]string) {
 	for key, data := range src {
 		dest[key] = data
 	}
-}
-
-// Fetch dependent data from validated data cache,
-// Returns the data and flag to indicate that if requested data 
-// is found in update request, the data should be merged with Redis data
-func (c *CVL) fetchDataFromRequestCache(tableName string, key string) (map[string]string, bool) {
-	cfgDataArr := c.requestCache[tableName][key]
-	if (cfgDataArr != nil) {
-		for _, cfgReqData := range cfgDataArr {
-			//Delete request doesn't have depedent data
-			if (cfgReqData.VOp == OP_CREATE) {
-				return cfgReqData.Data, false
-			} else	if (cfgReqData.VOp == OP_UPDATE) {
-				return cfgReqData.Data, true
-			}
-		}
-	}
-
-	return nil, false
-}
-
-//Fetch given table entries using pipeline
-func (c *CVL) fetchTableDataToTmpCache(tableName string, dbKeys map[string]interface{}) int {
-
-	TRACE_LOG(INFO_API, TRACE_CACHE, "\n%v, Entered fetchTableDataToTmpCache", time.Now())
-
-	totalCount := len(dbKeys)
-	if (totalCount == 0) {
-		//No entry to be fetched
-		return 0
-	}
-
-	entryFetched := 0
-	bulkCount := 0
-	bulkKeys := []string{}
-	for dbKey, val := range dbKeys { //for all keys
-
-		 if (val != nil) { //skip entry already fetched
-                        mapTable := c.tmpDbCache[tableName]
-                        delete(mapTable.(map[string]interface{}), dbKey) //delete entry already fetched
-                        totalCount = totalCount - 1
-                        if(bulkCount != totalCount) {
-                                //If some entries are remaining go back to 'for' loop
-                                continue
-                        }
-                } else {
-                        //Accumulate entries to be fetched
-                        bulkKeys = append(bulkKeys, dbKey)
-                        bulkCount = bulkCount + 1
-                }
-
-                if(bulkCount != totalCount) && ((bulkCount % MAX_BULK_ENTRIES_IN_PIPELINE) != 0) {
-                        //If some entries are remaining and bulk bucket is not filled,
-                        //go back to 'for' loop
-                        continue
-                }
-
-		mCmd := map[string]*redis.StringStringMapCmd{}
-
-		pipe := redisClient.Pipeline()
-
-		for _, dbKey := range bulkKeys {
-
-			redisKey := tableName + modelInfo.tableInfo[tableName].redisKeyDelim + dbKey
-			//Check in validated cache first and add as dependent data
-			if entry, mergeNeeded := c.fetchDataFromRequestCache(tableName, dbKey); (entry != nil) {
-				 c.tmpDbCache[tableName].(map[string]interface{})[dbKey] = entry 
-				 entryFetched = entryFetched + 1
-				 //Entry found in validated cache, so skip fetching from Redis
-				 //if merging is not required with Redis DB
-				 if (mergeNeeded == false) {
-					 continue
-				 }
-			}
-
-			//Otherwise fetch it from Redis
-			mCmd[dbKey] = pipe.HGetAll(redisKey) //write into pipeline
-			if mCmd[dbKey] == nil {
-				CVL_LOG(ERROR, "Failed pipe.HGetAll('%s')", redisKey)
-			}
-		}
-
-		_, err := pipe.Exec()
-		if err != nil {
-			CVL_LOG(ERROR, "Failed to fetch details for table %s", tableName)
-			return 0
-		}
-		pipe.Close()
-		bulkKeys = nil
-
-		mapTable := c.tmpDbCache[tableName]
-
-		for key, val := range mCmd {
-			res, err := val.Result()
-			if (err != nil || len(res) == 0) {
-				//no data found, don't keep blank entry
-				delete(mapTable.(map[string]interface{}), key)
-				continue
-			}
-			//exclude table name and delim
-			keyOnly := key
-
-			if (mapTable.(map[string]interface{})[keyOnly] != nil) {
-				tmpFieldMap := (mapTable.(map[string]interface{})[keyOnly]).(map[string]string)
-				//merge with validated cache data
-				mergeMap(res, tmpFieldMap)
-				fieldMap := c.checkFieldMap(&res)
-				mapTable.(map[string]interface{})[keyOnly] = fieldMap
-			} else {
-				fieldMap := c.checkFieldMap(&res)
-				mapTable.(map[string]interface{})[keyOnly] = fieldMap
-			}
-
-			entryFetched = entryFetched + 1
-		}
-
-		runtime.Gosched()
-	}
-
-	TRACE_LOG(INFO_API, TRACE_CACHE,"\n%v, Exiting fetchTableDataToTmpCache", time.Now())
-
-	return entryFetched
-}
-
-//populate redis data to cache
-func (c *CVL) fetchDataToTmpCache() *yparser.YParserNode {
-	TRACE_LOG(INFO_API, TRACE_CACHE, "\n%v, Entered fetchToTmpCache", time.Now())
-
-	entryToFetch := 0
-	var root *yparser.YParserNode = nil
-	var errObj yparser.YParserError
-
-	for entryToFetch = 1; entryToFetch > 0; { //Force to enter the loop for first time
-		//Repeat until all entries are fetched 
-		entryToFetch = 0
-		for tableName, dbKeys := range c.tmpDbCache { //for each table
-			entryToFetch = entryToFetch + c.fetchTableDataToTmpCache(tableName, dbKeys.(map[string]interface{}))
-		} //for each table
-
-		//If no table entry delete the table  itself
-		for tableName, dbKeys := range c.tmpDbCache { //for each table
-			if (len(dbKeys.(map[string]interface{}))  == 0) {
-				 delete(c.tmpDbCache, tableName)
-				 continue
-			}
-		}
-
-		if (entryToFetch == 0) {
-			//No more entry to fetch
-			break
-		}
-
-		if (Tracing == true) {
-			jsonDataBytes, _ := json.Marshal(c.tmpDbCache)
-			jsonData := string(jsonDataBytes)
-			TRACE_LOG(INFO_API, TRACE_CACHE, "Top Node=%v\n", jsonData)
-		}
-
-		data, err := jsonquery.ParseJsonMap(&c.tmpDbCache)
-
-		if (err != nil) {
-			return nil
-		}
-
-		//Build yang tree for each table and cache it
-		for jsonNode := data.FirstChild; jsonNode != nil; jsonNode=jsonNode.NextSibling {
-			TRACE_LOG(INFO_API, TRACE_CACHE, "Top Node=%v\n", jsonNode.Data)
-			//Visit each top level list in a loop for creating table data
-			topNode, _ := c.generateTableData(true, jsonNode)
-			if (root == nil) {
-				root = topNode
-			} else {
-				if root, errObj = c.yp.MergeSubtree(root, topNode); errObj.ErrCode != yparser.YP_SUCCESS {
-					return nil
-				}
-			}
-		}
-	} // until all dependent data is fetched
-
-	if root != nil && Tracing == true {
-		dumpStr := c.yp.NodeDump(root)
-		TRACE_LOG(INFO_DETAIL, TRACE_CACHE, "Dependent Data = %v\n", dumpStr)
-	}
-
-	TRACE_LOG(INFO_API, TRACE_CACHE, "\n%v, Exiting fetchToTmpCache", time.Now())
-	return root
-}
-
-
-func (c *CVL) clearTmpDbCache() {
-	for key, _ := range c.tmpDbCache {
-		delete(c.tmpDbCache, key)
-	}
-}
-
-func (c *CVL) generateTableFieldsData(config bool, tableName string, jsonNode *jsonquery.Node,
-parent *yparser.YParserNode) CVLRetCode {
-
-	//Traverse fields
-	for jsonFieldNode := jsonNode.FirstChild; jsonFieldNode!= nil;
-	jsonFieldNode = jsonFieldNode.NextSibling {
-		//Add fields as leaf to the list
-		if (jsonFieldNode.Type == jsonquery.ElementNode &&
-		jsonFieldNode.FirstChild != nil &&
-		jsonFieldNode.FirstChild.Type == jsonquery.TextNode) {
-
-			if (len(modelInfo.tableInfo[tableName].mapLeaf) == 2) {//mapping should have two leaf always
-				//Values should be stored inside another list as map table
-				listNode := c.addChildNode(tableName, parent, tableName) //Add the list to the top node
-				c.addChildLeaf(config, tableName,
-				listNode, modelInfo.tableInfo[tableName].mapLeaf[0],
-				jsonFieldNode.Data)
-
-				c.addChildLeaf(config, tableName,
-				listNode, modelInfo.tableInfo[tableName].mapLeaf[1],
-				jsonFieldNode.FirstChild.Data)
-
-			} else {
-				//check if it is hash-ref, then need to add only key from "TABLE|k1"
-				hashRefMatch := reHashRef.FindStringSubmatch(jsonFieldNode.FirstChild.Data)
-
-				if (hashRefMatch != nil && len(hashRefMatch) == 3) {
-				/*if (strings.HasPrefix(jsonFieldNode.FirstChild.Data, "[")) &&
-				(strings.HasSuffix(jsonFieldNode.FirstChild.Data, "]")) &&
-				(strings.Index(jsonFieldNode.FirstChild.Data, "|") > 0) {*/
-
-					c.addChildLeaf(config, tableName,
-					parent, jsonFieldNode.Data,
-					hashRefMatch[2]) //take hashref key value
-				} else {
-					c.addChildLeaf(config, tableName,
-					parent, jsonFieldNode.Data,
-					jsonFieldNode.FirstChild.Data)
-				}
-			}
-
-		} else if (jsonFieldNode.Type == jsonquery.ElementNode &&
-		jsonFieldNode.FirstChild != nil &&
-		jsonFieldNode.FirstChild.Type == jsonquery.ElementNode) {
-			//Array data e.g. VLAN members
-			for  arrayNode:=jsonFieldNode.FirstChild; arrayNode != nil;
-
-			arrayNode = arrayNode.NextSibling {
-				c.addChildLeaf(config, tableName,
-				parent, jsonFieldNode.Data,
-				arrayNode.FirstChild.Data)
-			}
-		}
-	}
-
-	return CVL_SUCCESS
-}
-
-func (c *CVL) generateTableData(config bool, jsonNode *jsonquery.Node)(*yparser.YParserNode, CVLErrorInfo) {
-	var cvlErrObj CVLErrorInfo
-
-	tableName := fmt.Sprintf("%s",jsonNode.Data)
-	c.batchLeaf = ""
-
-	//Every Redis table is mapped as list within a container,
-	//E.g. ACL_RULE is mapped as 
-	// container ACL_RULE { list ACL_RULE_LIST {} }
-	var topNode *yparser.YParserNode
-
-	// Add top most conatiner e.g. 'container sonic-acl {...}'
-	if _, exists := modelInfo.tableInfo[tableName]; exists == false {
-		return nil, cvlErrObj 
-	}
-	topNode = c.yp.AddChildNode(modelInfo.tableInfo[tableName].module,
-	nil, modelInfo.tableInfo[tableName].modelName)
-
-	//Add the container node for each list 
-	//e.g. 'container ACL_TABLE { list ACL_TABLE_LIST ...}
-	listConatinerNode := c.yp.AddChildNode(modelInfo.tableInfo[tableName].module,
-	topNode, tableName)
-
-	//Traverse each key instance
-	for jsonNode = jsonNode.FirstChild; jsonNode != nil; jsonNode = jsonNode.NextSibling {
-
-		//For each field check if is key 
-		//If it is key, create list as child of top container
-		// Get all key name/value pairs
-		if yangListName := getRedisKeyToYangList(tableName, jsonNode.Data); yangListName!= "" {
-			tableName = yangListName
-		}
-		keyValuePair := getRedisToYangKeys(tableName, jsonNode.Data)
-		keyCompCount := len(keyValuePair)
-		totalKeyComb := 1
-		var keyIndices []int
-
-		//Find number of all key combinations
-		//Each key can have one or more key values, which results in nk1 * nk2 * nk2 combinations
-		idx := 0
-		for i,_ := range keyValuePair {
-			totalKeyComb = totalKeyComb * len(keyValuePair[i].values)
-			keyIndices = append(keyIndices, 0)
-		}
-
-		for  ; totalKeyComb > 0 ; totalKeyComb-- {
-			//Get the YANG list name from Redis table name
-			//Ideally they are same except when one Redis table is split
-			//into multiple YANG lists
-
-			//Add table i.e. create list element
-			listNode := c.addChildNode(tableName, listConatinerNode, tableName + "_LIST") //Add the list to the top node
-
-			//For each key combination
-			//Add keys as leaf to the list
-			for idx = 0; idx < keyCompCount; idx++ {
-				c.addChildLeaf(config, tableName,
-				listNode, keyValuePair[idx].key,
-				keyValuePair[idx].values[keyIndices[idx]])
-			}
-
-			//Get all fields under the key field and add them as children of the list
-			c.generateTableFieldsData(config, tableName, jsonNode, listNode)
-
-			//Check which key elements left after current key element
-			var next int = keyCompCount - 1
-			for  ((next > 0) && ((keyIndices[next] +1) >=  len(keyValuePair[next].values))) {
-				next--
-			}
-			//No more combination possible
-			if (next < 0) {
-				break
-			}
-
-			keyIndices[next]++
-
-			//Reset indices for all other key elements
-			for idx = next+1;  idx < keyCompCount; idx++ {
-				keyIndices[idx] = 0
-			}
-
-			TRACE_LOG(INFO_API, TRACE_CACHE, "Starting batch leaf creation - %s\n", c.batchLeaf)
-			//process batch leaf creation
-			if errObj := c.yp.AddMultiLeafNodes(modelInfo.tableInfo[tableName].module, listNode, c.batchLeaf); errObj.ErrCode != yparser.YP_SUCCESS {
-				cvlErrObj = CreateCVLErrObj(errObj)
-				return nil, cvlErrObj 
-			}
-			c.batchLeaf = ""
-		}
-	}
-
-	return topNode, cvlErrObj
 }
 
 func (c *CVL) translateToYang(jsonMap *map[string]interface{}) (*yparser.YParserNode, CVLErrorInfo) {
@@ -1585,45 +1166,6 @@ func (c *CVL) validateSyntax(data *yparser.YParserNode) (CVLErrorInfo, CVLRetCod
 	}
 
 	return cvlErrObj, CVL_SUCCESS
-}
-
-//Perform semantic checks 
-func (c *CVL) validateSemantics(data *yparser.YParserNode, appDepData *yparser.YParserNode) (CVLErrorInfo, CVLRetCode) {
-	var cvlErrObj CVLErrorInfo
-	
-	if (SkipSemanticValidation() == true) {
-		return cvlErrObj, CVL_SUCCESS
-	}
-
-	//Get dependent data from 
-	depData := c.fetchDataToTmpCache() //fetch data to temp cache for temporary validation
-
-	if (Tracing == true) {
-		TRACE_LOG(INFO_API, TRACE_SEMANTIC, "Validating semantics data=%s\n depData =%s\n, appDepData=%s\n....", c.yp.NodeDump(data), c.yp.NodeDump(depData), c.yp.NodeDump(appDepData))
-	}
-
-	if errObj := c.yp.ValidateSemantics(data, depData, appDepData); errObj.ErrCode != yparser.YP_SUCCESS {
-
-		retCode := CVLRetCode(errObj.ErrCode)
-
-		cvlErrObj =  CVLErrorInfo {
-			TableName : errObj.TableName,
-		        ErrCode   : CVLRetCode(errObj.ErrCode),		
-			CVLErrDetails : cvlErrorMap[retCode], 
-			Keys      : errObj.Keys,
-			Value     : errObj.Value,
-			Field     : errObj.Field,
-			Msg       : errObj.Msg,
-			ConstraintErrMsg : errObj.ErrTxt,
-			ErrAppTag	: errObj.ErrAppTag,
-		}
-
-
-
-		return  cvlErrObj, retCode
-	}
-
-	return cvlErrObj ,CVL_SUCCESS
 }
 
 //Add config data item to accumulate per table
