+++ cvl/schema/Makefile
@@ -28,19 +28,27 @@ out_dir=$(TOPDIR)/build/cvl/schema
 
 src_files=$(wildcard $(sonic_yang)/*.yang)
 src_files += $(wildcard $(sonic_yang_common)/*.yang)
-out=$(patsubst %.yang, $(out_dir)/%.yin, $(notdir $(src_files)))
-search_path=$(std_yang_common):$(sonic_yang):$(sonic_yang_common)
+out=$(patsubst %.yang, $(out_dir)/%.yin, $(shell ls -1 $(sonic_yang)/*.yang | cut -d'/' -f6)) 
+out_common=$(patsubst %.yang, $(out_dir)/%.yin, $(shell ls -1 $(sonic_yang_common)/*.yang | cut -d'/' -f7)) 
+out_platform=$(patsubst %.yang, $(out_dir)/%.yin, $(shell find $(sonic_yang_platform) -name '*.yang' | cut -d'/' -f6-8))
+out_platform_dep=$(shell find $(sonic_yang_platform) -name '*.yang')
+out_tree=$(patsubst %.yang, $(out_dir)/%.tree, $(src_files)) 
 
-all: schema
+search_path=$(sonic_yang):$(sonic_yang_common):$(sonic_yang_common)/ietf
 
-schema: $(out)
 
-.PRECIOUS: %/.
-%/.:
-	mkdir -p $(@D)
+all: precheck schema
+
+precheck:
+	mkdir -p $(out_dir) 
+
+schema: $(out) $(out_common) $(out_platform)
+#	@$(call install_cvl_schema)
+
+schema-tree: $(out_tree)
 
 #Build YANG models
-$(out_dir)/%.yin:$(sonic_yang)/%.yang | $(out_dir)/.
+$(out_dir)/%.yin:$(sonic_yang)/%.yang
 	@echo "Generating `basename $@` ..."
 	@devFile="`echo $@ | cut -d . -f1`-deviation.yang"; \
 	if [ -f $$devFile ] ; then devOpt="--deviation-module $$devFile"; fi; \
@@ -49,12 +57,28 @@ $(out_dir)/%.yin:$(sonic_yang)/%.yang | $(out_dir)/.
 
 
 #Build common YANG models
-$(out_dir)/%.yin:$(sonic_yang_common)/%.yang | $(out_dir)/.
+$(out_dir)/%.yin:$(sonic_yang_common)/%.yang
 	@echo "Generating `basename $@` ..."
 	@devFile="`echo $@ | cut -d . -f1`-deviation.yang"; \
 	if [ -f $$devFile ] ; then devOpt="--deviation-module $$devFile"; fi; \
 	pyang -p $(search_path)	--plugindir $(pyang_plugin_dir) \
 		-f yin-cvl $$devOpt $< -o $@
 
+#Build platform specific YANG models
+$(out_platform):$(out_platform_dep)
+	@mkdir -p `dirname $@`
+	@echo "Generating $@ ..."
+	@pyang -p $(search_path) --plugindir $(pyang_plugin_dir) \
+		-f yin-cvl "$(sonic_yang)/`echo $@ | grep -o 'platform/.*\.'`yang" -o $@
+
+$(out_dir)/%.tree:%.yang
+	@echo "Generating `basename $@` ..."
+	@devFile="`echo $< | cut -d . -f1`-dev.yang"; \
+	if [ -f $$devFile ] ; then devOpt="--deviation-module $$devFile"; fi; \
+	pyang -p $(search_path) -f tree $$devOpt $< -o `basename $@`
+
 clean:
-	$(RM) -r $(out_dir)
+	@echo "Removing files ..."
+	rm -rf $(out_dir)
+	rm -rf platform/
+
