+++ models/yang/sonic/common/sonic-common.yang
@@ -32,6 +32,14 @@ module sonic-common {
 		}
 	}
 
+	typedef oper-status {
+		type enumeration {
+			enum up;
+			enum down;
+		}
+	}
+
+
 	container operation {
 		description "This definition is used internally by CVL and 
 			is not exposed in NBI. Leaf 'operation' allows
